bawarchi
All tasks for the project have been fully completed. gemini API was integrated to build a working chat interface. Improvements in the UI have been added for code snippets, which have the functionality of syntax highlighting, copy-to-clipboard proper formatting. Support for file uploads was implemented, providing the possibility for validation and then the integration of the file's content into chat. The voice chat feature, speech-to-text, had been implemented into the system.  User authentication was implemented with secure registration, login, and session management, and chat history preservation was enabled by integrating the database, while all features were deployed on the project ensuring that everything worked correctly except for the result data streaming whose backend code was written but issues occurred in the React frontend.